replication: unifying the babu db interface, by rebuilding the bypass to the replication
git-svn-id: https://babudb.googlecode.com/svn/trunk@128 d4c7dd96-b7b2-11dd-893d-d98698f66bbf

+++ java/src/org/xtreemfs/babudb/sandbox/ContinuesRandomGenerator.java
@@ -49,7 +49,7 @@ public class ContinuesRandomGenerator {
     public final static int MAX_VALUE_LENGTH = 30;
     public final static int MIN_VALUE_LENGTH = 10;
     
-    public final static long MAX_SEQUENCENO = Long.MAX_VALUE;
+    public final static long MAX_SEQUENCENO = Integer.MAX_VALUE;
     public final static long MIN_SEQUENCENO = BabuDBLongrunTestConfig.MIN_SEQUENCENO;
 		
     private final static byte[] CHARS;
